// frontend/src/app/models/master/masGeneralSkillData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

import { MasGeneralSkillLevelData } from "./masGeneralSkillLevelData";

export interface MasGeneralSkillData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    first?: number;
    /** รหัสทักษะทั่วไป */
    generalSkillCode?: string;
    /** pk ทักษะทั่วไป */
    generalSkillId?: number;
    /** ชื่อทักษะทั่วไป (อังกฤษ) */
    generalSkillNameEn: string;
    /** ชื่อทักษะทั่วไป (ไทย) */
    generalSkillNameTh: string;

    mode?: string;
    rowNum?: number;
    size?: number;

    /**add */
    skillImage?: string;
    filename?: string;
    module?: number;
    prefix?: string;
    /** ประเภทหมวดหมู่ LOOKUP (30021000)  */
    skillFormat?: string;
    /** custom รูปคน */
    base64?: string;

    /** custom Fk(mas_general_skill_level) */
    skillLevelList?: MasGeneralSkillLevelData[];
}